managedResources/requirements.txt
fastapi==0.104.1
influxdb-client==1.39.0
numpy==1.26.2
paho-mqtt==1.6.1
pandas==2.1.4
uvicorn==0.24.0.post1
uvloop==0.19.0
websockets==12.0
